skillswap/backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg[binary,pool]==3.2.1
redis==5.0.7
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
jinja2==3.1.4
